Ignore offGrid result
The resulting feeder.offGrid value does not match what we asked for,
even though it does (eventually) update correctly.

We could wait until we see the right status, but might as well just
assume it's going to work if we got back a non-error response, and let
the component refresh the device status.