Handle each sighting independently
A single SightingReport may contain multiple sightings: some of which
can be finished as-is, some may need a manual species selection, and
some may need to be collected as a mystery visitor.

Depending on the chosen postcard finishing strategy, we can handle each
sighting separately - i.e., taking the recognized sightings and choosing
the highest-confidence match if possible, but discarding unrecognized
species.