Update schema
This introduces crowd-sourced media identification requests support.
The user will be able to request community help with identifying a bird
species. When enough users have voted and there seems to be a clear
winning consensus, the identification is accepted. This works similarly
to the app's ability to send a photo to BB's experts to classify, but
leveraging the community.

Eventually we may be able to add support for this into the collection
automation, where a low-confidence identification can be submitted for
community identification, rather than accepting it as a "mystery"
visitor.